refactor: move error types to the error file
Replace ProcessErrorResults with errors.Join which does the same thing
for us.
