+++ internal/juju/utils.go
@@ -152,11 +152,3 @@ func WaitForAppsAvailable(ctx context.Context, client *apiapplication.Client, ap
 		}
 	}
 }
-
-// ProcessErrorResults processes the results of a secret operation.
-func ProcessErrorResults(results []error) error {
-	if results[0] != nil && len(results) > 1 {
-		return &MultiError{Errors: results}
-	}
-	return nil
-}
